requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.1
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
backoff==2.2.1
certifi==2023.5.7
charset-normalizer==3.2.0
chromadb==0.3.26
click==8.1.4
clickhouse-connect==0.6.6
coloredlogs==15.0.1
comm==0.1.3
dataclasses-json==0.5.9
debugpy==1.6.7
decorator==5.1.1
dnspython==2.3.0
duckdb==0.8.1
exceptiongroup==1.1.2
executing==1.2.0
fastapi==0.100.0
flatbuffers==23.5.26
frozenlist==1.3.3
greenlet==2.0.2
h11==0.14.0
hnswlib==0.7.0
httptools==0.6.0
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
ipykernel==6.24.0
ipython==8.14.0
ipywidgets==8.0.7
jedi==0.18.2
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyterlab-widgets==3.0.8
langchain==0.0.228
langchainplus-sdk==0.0.20
loguru==0.7.0
lz4==4.3.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib-inline==0.1.6
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.6
numexpr==2.8.4
numpy==1.25.1
onnxruntime==1.15.1
openai==0.27.8
openapi-schema-pydantic==1.2.4
overrides==7.3.1
packaging==23.1
pandas==2.0.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pinecone-client==2.2.2
platformdirs==3.8.1
posthog==3.0.1
prompt-toolkit==3.0.39
protobuf==4.23.4
psutil==5.9.5
ptyprocess==0.7.0
pulsar-client==3.2.0
pure-eval==0.2.2
pydantic==1.10.11
Pygments==2.15.1
PyPDF2==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
pyzmq==25.1.0
requests==2.31.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.18
stack-data==0.6.2
starlette==0.27.0
sympy==1.12
tenacity==8.2.2
tokenizers==0.13.3
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.3
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
wcwidth==0.2.6
websockets==11.0.3
widgetsnbextension==4.0.8
yarl==1.9.2
zipp==3.16.0
zstandard==0.21.0
